requirements.txt
annotated-types==0.5.0
arabic-reshaper==3.0.0
asgiref==3.2.10
asn1crypto==1.5.1
bson==0.5.8
cart==1.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
crypto==1.4.1
cryptography==41.0.3
cssselect2==0.7.0
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.28
django-admin-rangefilter==0.11.1
django-betterforms==1.2
django-cart==1.0.4
django-ckeditor==5.9.0
django-coupon-management==1.2
django-crispy-forms==1.9.1
django-extensions==3.0.2
django-js-asset==1.2.2
django-mapbox-location-field==1.5.0
django-mathfilters==1.0.0
django-qr-code==3.0.0
django-widget-tweaks==1.4.8
djongo==1.3.2
dnspython==1.16.0
docutils==0.20.1
flake8==6.1.0
graphviz==0.14
html5lib==1.1
idna==3.4
Kivy==2.2.1
Kivy-Garden==0.1.5
kivy-garden.kivymd==2019.910
kivy-garden.xcamera==2020.613
kivy-garden.zbarcam==2020.524
lxml==4.9.3
mccabe==0.7.0
mysql-connector-python==8.0.20
Naked==0.1.32
numpy==1.25.2
opencv-python==4.8.0.76
oscrypto==1.3.0
Pillow==10.0.1
protobuf==3.12.2
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
pydot==1.4.1
pyflakes==3.1.0
Pygments==2.16.1
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pymongo==3.10.1
pyparsing==2.4.7
pypdf==3.16.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
QrScan==2020.528
raven==6.10.0
reportlab==3.6.13
requests==2.31.0
segno==1.5.2
shellescape==3.8.1
six==1.15.0
sqlparse==0.2.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
validators==0.22.0
webencodings==0.5.1
whitenoise==6.6.0
xcamera==2020.613
xhtml2pdf==0.2.11
zbarcam==2020.524
zxing==1.0
